docusmart-backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# OCR
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
opencv-python-headless==4.9.0.80

# AI / NLP
anthropic==0.18.1
spacy==3.7.4
openai==1.12.0

# PDF Generation
reportlab==4.1.0
weasyprint==61.2

# Email
sendgrid==6.11.0

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.27.0

# Development
pytest==8.0.1
pytest-asyncio==0.23.5
